python3-01/requirements.txt
fastapi
uvicorn
mysql-connector-python
mysqlclient
